CK 150306  - Finalised migration of portable library with full support with required changes to suit each platform  - Changes to Ject to be compatible with PCL (passed all test cases)  - Also added regression models and started adding new classifier functionality

+++ numl/Supervised/NeuralNetwork/Network.cs
@@ -40,7 +40,7 @@ namespace numl.Supervised.NeuralNetwork
             IFunction ident = new Ident();
 
             // set number of hidden units to (Input + Hidden) * 2/3 as basic best guess. 
-            int hidden = (int)System.Math.Ceiling((decimal)(x.Cols + output) * 2m / 3m);
+            int hidden = (int)System.Math.Ceiling((double)(x.Cols + output) * 2.0 / 3.0);
 
             // creating input nodes
             nn.In = new Node[x.Cols + 1];
